Improve folder structure. Moved more config stuff into index and then broke each component into its own file.
