+++ src/components/index.ts
@@ -1,3 +1,6 @@
+export { default as Home } from './Home';
+export { default as Login } from './Login';
 export { default as MyComp } from './MyComp';
+export { default as PrivateRoute } from './PrivateRoute';
 export { default as Tray } from './Tray';
 export { default as TrayItem } from './TrayItem';
